+++ apps/autotest/models/TST/_initialData/010_create_navshortcuts.js
@@ -24,6 +24,7 @@ module.exports = function (session) {
       execParams: {
         code: 'tst_desktop',
         caption: 'Test',
+        displayOrder: 9000,
         iconCls: 'u-icon-desktop-document-types'
       }
     })
@@ -316,25 +317,25 @@ module.exports = function (session) {
     }
   })
 
-  displayOrder += 10
-  console.log('\t\t\tcreate `tst_aclrls` shortcut')
-  conn.insert({
-    fieldList: ['ID'],
-    entity: 'ubm_navshortcut',
-    execParams: {
-      desktopID: desktopID,
-      code: 'tst_aclrls',
-      caption: 'tst_aclrls',
-      displayOrder: displayOrder,
-      cmdCode: JSON.stringify({
-        renderer: 'vue',
-        cmdType: 'showList',
-        cmdData: {
-          entityName: 'tst_aclrls'
-        }
-      }, null, '\t')
-    }
-  })
+  // displayOrder += 10
+  // console.log('\t\t\tcreate `tst_aclrls` shortcut')
+  // conn.insert({
+  //   fieldList: ['ID'],
+  //   entity: 'ubm_navshortcut',
+  //   execParams: {
+  //     desktopID: desktopID,
+  //     code: 'tst_aclrls',
+  //     caption: 'tst_aclrls',
+  //     displayOrder: displayOrder,
+  //     cmdCode: JSON.stringify({
+  //       renderer: 'vue',
+  //       cmdType: 'showList',
+  //       cmdData: {
+  //         entityName: 'tst_aclrls'
+  //       }
+  //     }, null, '\t')
+  //   }
+  // })
 
   displayOrder += 10
   console.log('\t\t\tcreate `tst_gridAction` shortcut')