+++ packages/ubcli/lib/migrate.js
@@ -29,7 +29,7 @@ const base = require('@unitybase/base')
 const options = base.options
 const argv = base.argv
 const execSql = require('./execSql')
-const { createDBConnectionPool, releaseDBConnectionPool } = require('@unitybase/base')
+const { createDBConnectionPool, releaseDBConnectionPool, GC_KEYS } = require('@unitybase/base')
 const generateDDL = require('./generateDDL')
 const { normalizeVersion, updateVersionsInDB } = require('./flow/migrationUtils')
 
@@ -86,8 +86,8 @@ module.exports = function migrate (cfg) {
         help: 'Output execution time for each command into console'
       })
     cfg = opts.parseVerbose({}, true)
-    if (cfg.verbose) cfg.verboseShort = true
     if (!cfg) return
+    if (cfg.verbose) cfg.verboseShort = true
   }
   cfg.user = 'root'
   // increase receive timeout to 10 minutes - in case DB server is slow we can easily reach 30s timeout
@@ -106,10 +106,13 @@ const NORMALIZE_VERSION_RE = /^((\d{1,3})[._](\d{1,3})[._](\d{1,3}))/
  *  @private
  */
 function runMigrations (params) {
+  if (process.globalCachePut) { // UB@5.22.10
+    process.globalCachePut(GC_KEYS.UBQ_SCHEDULER_INITIALIZED, 'yes')
+  }
   const serverConfig = argv.getServerConfiguration(false)
   let dbConnections = createDBConnectionPool(serverConfig.application.connections)
-  const multitenancyEnabled = serverConfig.security.multitenancy
-    && serverConfig.security.multitenancy.enabled
+  const multitenancyEnabled = serverConfig.security.multitenancy &&
+    serverConfig.security.multitenancy.enabled
   if (params.tenantID && !multitenancyEnabled) {
     throw new Error('tenantID parameter should only be used when multitenancy enabled')
   }